refactor: reorganize the folders components from the task and tasks-list
