+++ web/src/components/task-list/components/task/index.ts
@@ -1,0 +1,1 @@
+export * from "@/components/task-list/components/task/task";
