--- web/src/components/task/index.ts
@@ -1,1 +1,0 @@
-export * from "@/components/task/task";
